lib/vdso: Add vdso/vvardef.h and move x86 vgetrandom definitions
The vdso/vsyscall.h is currently only included in the time/vsyscall.c
generic code, so the DEFINE_VVAR/DEFINE_VVAR_SINGLE are only defined
once.  To allow the header to be included multiple times, add a new
generic header, vvardef.h, and move the x86 definitions to it.

This is to allow the parametrize of the _vdso_rng_data access similar
to vdso_data.

Signed-off-by: Adhemerval Zanella <adhemerval.zanella@linaro.org>